scrittura file

+++ src/main/MetztliTeam.java
@@ -3,17 +3,17 @@ package main;
 import java.util.ArrayList;
 
 public class MetztliTeam {
-	private static int costoPercorso;
+	private static double costoPercorso;
 	private static ArrayList<Integer> percorso;
 	public static final String DESCRIZIONE_TEAM=
 			"Il veicolo di questo team consuma carburante in base all'altitudine percorsa";
 	private static final String NOME_TEAM="Metzli";
 
-	public static int getCostoPercorso() {
+	public static double getCostoPercorso() {
 		return costoPercorso;
 	}
 
-	public static void setCostoPercorso(int costoPercorso) {
+	public static void setCostoPercorso(double costoPercorso) {
 		MetztliTeam.costoPercorso = costoPercorso;
 	}
